backend/requirements.txt
fastapi==0.78.0
uvicorn==0.23.2
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==1.10.13
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7
email-validator==2.1.0
